removed maya_lib2 and moved to maya_lib
