--- python/vtool/maya_lib2/__init__.py
@@ -1,1 +1,0 @@
-# Copyright (C) 2022 Louis Vottero louis.vot@gmail.com    All rights reserved.
